pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "game-theory-ai"
version = "0.1.0"
description = "博弈AI分析工具"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "GameTheoryAI Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Topic :: Games/Entertainment"
]
dependencies = [
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "flask>=2.0.0",
    "psutil>=5.9.0",
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "pytesseract>=0.3.0",
    "tencentcloud-sdk-python>=3.0.0"
]

[project.optional-dependencies]
dev = [
    "setuptools>=45",
    "wheel>=0.38.0",
    "setuptools_scm>=6.2",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0"
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools_scm]
write_to = "src/_version.py"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203"]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_config = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 